+++ resources/views/emails/member-approved-notification.blade.php
@@ -9,6 +9,12 @@ Ihr Antrag auf Mitgliedschaft bei der DITIB Ahlen wurde angenommen.
 
 Willkommen in unserer Gemeinde.
 
+Sie können Ihre Mitgliedsdaten jederzeit einsehen und bearbeiten:
+
+<x-mail::button url="https://mitglied.ditib-ahlen-projekte.de/konto">
+Zum Mitgliedskonto
+</x-mail::button>
+
 Mit freundlichen Grüßen,<br>
 Ihr Team von {{ config('app.name') }}
 </x-mail::message>